Show changelog before installing update

+++ app/src/main/java/io/xpipe/app/core/AppCache.java
@@ -34,6 +34,10 @@ public class AppCache {
     }
 
     public static void clear() {
+        if (!Files.exists(getBasePath())) {
+            return;
+        }
+
         try {
             FileUtils.cleanDirectory(getBasePath().toFile());
         } catch (IOException e) {
